Revert postgres plugin to use gzip and gunzip
While the change introduced in #115 achieves role backups, the switch from
the gzip/gunzip tool to tar introduces file boundary artifacts in the
backup stream, which psql is unable to handle. By using gzip/gunzip in lieu
of tar, we are able to concatenate and preserve equivalent behavior, but
without the extraneous file boundary markers.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>